Develop orges

+++ src/main/java/com/internship/HRapp/service/concretes/RoleServiceImpl.java
@@ -20,7 +20,13 @@ public class RoleServiceImpl implements RoleServiceInterface {
 
     @Override
     public RoleDTO getRoleById(UUID roleId) {
-    return roleMapper.toDTO(rolesRepo.getById(roleId));
+        boolean exists = rolesRepo.existsById(roleId);
+        if (!exists){
+            throw new IllegalStateException(
+                    "Role with id " + roleId + " does not exist!"
+            );
+        }
+        return roleMapper.toDTO(rolesRepo.getById(roleId));
     }
     @Override
     public RoleDTO getRoleByRoleName(String roleName) {
@@ -39,6 +45,12 @@ public class RoleServiceImpl implements RoleServiceInterface {
     }
     @Override
     public String deleteRolesById(UUID roleId) {
+        boolean exists = rolesRepo.existsById(roleId);
+        if (!exists){
+            throw new IllegalStateException(
+                    "Role with id " + roleId + " does not exist!"
+            );
+        }
         rolesRepo.deleteById(roleId);
         return "role removed {}" + roleId;
     }
